refactor: remove fromScratch
